Fix issue with copying gif url to clipboard
BOOL test = [[NSPasteboard generalPasteboard] writeObjects:@[self.gif.downloadURL]];
is returning YES but the url can't be pasted later into a text editor.
Going to Finder > Edit > Show Clipboard says the clipboard contents
are unknown.

Seems like the system is writing the url as a plist?
Switching to write out the absoluteString of the url works for pasting.